Moving q2 types into respective modules
